+++ CHANGELOG.md
@@ -8,6 +8,8 @@ https://github.com/coleifer/peewee/releases
 ## master
 
 * Add bitwise and other helper methods to `BigBitField`, #2802.
+* Add `add_column_default` and `drop_column_default` migrator methods for
+  specifying a server-side default value, #2803.
 * The new `star` attribute was causing issues for users who had a field named
   star on their models. This attribute is now renamed to `__star__`. #2796.